Fixes warnings and adds flag to treat warnings as errors

+++ include/rendering/buffers/IndexData.h
@@ -4,6 +4,6 @@
 
 namespace erm {
 	
-	typedef unsigned short IndexData;
+	typedef unsigned int IndexData;
 	
 }
